Split DSL into ResourceDSL and PageDSL
