+++ spec/unit/resource_registration_spec.rb
